airflow/requirements.txt
apache-airflow-providers-google
apache-airflow-providers-http
pyarrow
pandas
sqlalchemy
psycopg2-binary
pyspark